pyproject.toml
[project]
name = "gf_25_v3"
version = "0.1.0"
description = "GF-25 v3 FastAPI service with Ollama optimizations"
authors = [ { name = "GF-25" } ]
requires-python = ">=3.11"
readme = "README.md"

# Core dependencies (pinned per [DEPS-CORE])
dependencies = [
    "pydantic>=2.5.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "ollama>=0.3.0",
    "litellm>=1.42.0",
    "tenacity>=9.1.2",
    "prometheus-client>=0.22.1",
    "pyyaml>=6.0.1",
    "watchfiles>=0.21.0",
    "asgi-lifespan>=2.1.0",
]

[project.optional-dependencies]
# Developer tools
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "anyio>=4.0.0",
]

[tool.uv]
# uv settings can stay default; this enables `uv sync` from pyproject

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "anyio>=4.10.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.8",
]
